Altered configuration format of transfer learning to only contain multi_source entries.
The list of stages before the list of multi_source tasks in the config file was superfluous, since the standalone list of stages was treated identically to a multi_source stage with iteration = 1 (which is the same as omitting the iteration value).  So I've removed this functionality, simplifying both the configuration format and the code.

Also removed an unused imported module from stage.py.